% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_chrm functions.R
\name{createdata}
\alias{createdata}
\title{Cluster multinomial data}
\usage{
createdata(p1_a, E_a, s2_a, meanvalues, sizes)
}
\arguments{
\item{p1_a}{the between outcome correlation value}

\item{E_a}{the within outcome correlation value}

\item{s2_a}{a vector of variances for each visit for both outcome combined.}

\item{meanvalues}{Matrix containing the means for each visit cols for each cluster rows. this should have the same number of cols as the length of s2_a}

\item{sizes}{avector with the sizes for each cluster. This should have the same number of rows as the matrix meanvalues}
}
\value{
a matrix with data geneterated using
}
\description{
Cluster multinomial data
}
\examples{
createdata(p1_a=.5, E_a=c(.3,.4,.6,.4) ,s2_a=c(1,2,2,1,1,1,2,2), meanvalues=matrix(c(3,3,3,6,2,2,6,2,2,6,5,2,3,3,3,7,5,5,8,4,4,9,7,2), ncol=8), sizes=c(100,100,100))
}
