% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_chrm functions.R
\name{find_optimal_clusters_single3d}
\alias{find_optimal_clusters_single3d}
\title{Title Clustering by Spectral decomposition}
\usage{
find_optimal_clusters_single3d(
  arr,
  id,
  timevector,
  B,
  cor1,
  var_a = TRUE,
  cor_o = TRUE,
  cor_t = TRUE,
  method = "scaledPCA",
  algo = "kmeans",
  dis = "Cholesky",
  gapdis = "euclidean",
  gapmethod = "Tibs2001SEmax",
  covmethod = "pearson",
  standard = TRUE,
  fixedk = FALSE
)
}
\arguments{
\item{arr}{An array with the continuous variables of interest}

\item{id}{A vector of IDs associated with each individual in the order of the matrix}

\item{timevector}{The visits or time points associated with each column for each variable. Example: timevector = C(0,1,2,3)}

\item{B}{the number of uniform samples to be drawn for the GAP statistics}

\item{cor1}{A square matrix with the correlations between the outcomes}

\item{var_a}{a boolean indicator of whether to account for the variances in the diagonal matrix}

\item{method}{the version of the Tibshirani Gap statistic to use "original" or "scaledPCA"}

\item{algo}{what algorithm to use to partition clusters: "kmeans", "kmedians", "kmeans2", "kmedians2", "pam", etc.... Note that kmeans2 and kmedians2 performs kmeans and kmedians 20 times with different starting points and selects the cluster partition with the best objective function.}

\item{dis}{the distance to use ("euclidean","manhattan" , "mahalanobis","minkowski", "maximum", "canberra", "binary","mahalanobis-ZCA", "mahalanobis-ZCA-cor","mahalanobis-PCA", "mahalanobis-PCA-cor", "mahalanobis-Cholesky")}

\item{gapdis}{the distance to use in the gap statistic function. Same options as "dis". Keep as euclidean if "dis" is already non-euclidean.}

\item{gapmethod}{which method to use to obtain optimal clusters ("Tibs2001SEmax", "globalSEmax", "fixed", "RamdomForrest")}

\item{covmethod}{Correlation method. "pearson" or "spearman".}

\item{standard}{Should each variable be globally standardized recursively before running algorithm at each iteration}

\item{fixedk}{The number of clusters to produce if you would like to fix the number of clusters. Otherwise, keep as FALSE
#'}

\item{cor_0}{a boolean indicator of whether to account for the correlations between the outcomes (obtained from cor1)}

\item{cor_e}{a boolean indicator of whether to account for the temporal correlations for the repeated measures}
}
\value{
returns a list of data frames containing ID, repeated variable 1, cluster partitions at each time point.
}
\description{
Title Clustering by Spectral decomposition
}
\examples{
find_optimal_clusters_single3d(arr=myarray, id=idvector, B=500, cor1=cor(myarray[[1]][,1],myarray[[2]][,1]) method="scaledPCA", algo="kmeans", dis="Cholesky", gapdis="euclidean", gapmethod="Tibs2001SEmax", covmethod="pearson", var_a = TRUE, cor_o=TRUE, cor_t=TRUE, novarcor=FALSE,standard=FALSE, fixmat=FALSE)
}
